+++ sdk/communication/azure-communication-phonenumbers/src/main/java/com/azure/communication/phonenumbers/implementation/models/PhoneNumberSearchRequest.java
@@ -4,6 +4,7 @@
 
 package com.azure.communication.phonenumbers.implementation.models;
 
+import com.azure.communication.phonenumbers.CodeCoverageAnnotation.Generated;
 import com.azure.communication.phonenumbers.models.PhoneNumberAssignmentType;
 import com.azure.communication.phonenumbers.models.PhoneNumberCapabilities;
 import com.azure.communication.phonenumbers.models.PhoneNumberType;
@@ -15,6 +16,7 @@ import com.fasterxml.jackson.annotation.JsonProperty;
  * following purchase.
  */
 @Fluent
+@Generated
 public final class PhoneNumberSearchRequest {
     /*
      * The type of phone numbers to search for, e.g. geographic, or tollFree.