+++ sdk/communication/azure-communication-phonenumbers/src/main/java/com/azure/communication/phonenumbers/models/PhoneNumberAreaCode.java
@@ -7,7 +7,9 @@ package com.azure.communication.phonenumbers.models;
 import com.azure.core.annotation.Immutable;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
-/** Represents an Area Code. */
+/**
+ * The PhoneNumberAreaCode model.
+ */
 @Immutable
 public final class PhoneNumberAreaCode {
     /*