+++ sdk/cosmos/azure-cosmos-spark_3_2-12/src/test/scala/com/azure/cosmos/spark/SparkE2EWriteITest.scala
@@ -7,7 +7,7 @@ import com.azure.cosmos.spark.CosmosPatchOperationTypes.CosmosPatchOperationType
 import com.azure.cosmos.spark.ItemWriteStrategy.ItemWriteStrategy
 import com.azure.cosmos.spark.diagnostics.BasicLoggingTrait
 import org.apache.spark.sql.functions.{col, from_json}
-import org.apache.spark.sql.types.{MapType, StringType, StructField, StructType}
+import org.apache.spark.sql.types.{StringType, StructField, StructType}
 
 class SparkE2EWriteITest
   extends IntegrationSpec
@@ -22,26 +22,42 @@ class SparkE2EWriteITest
   //scalastyle:off magic.number
   //scalastyle:off null
 
-  private case class UpsertParameterTest(bulkEnabled: Boolean, itemWriteStrategy: ItemWriteStrategy, hasId: Boolean = true)
+  private case class UpsertParameterTest(bulkEnabled: Boolean, itemWriteStrategy: ItemWriteStrategy, hasId: Boolean = true, initialBatchSize: Option[Int] = None)
 
   private val upsertParameterTest = Seq(
-    UpsertParameterTest(bulkEnabled = true, itemWriteStrategy = ItemWriteStrategy.ItemOverwrite),
-
-    UpsertParameterTest(bulkEnabled = false, itemWriteStrategy = ItemWriteStrategy.ItemOverwrite),
-    UpsertParameterTest(bulkEnabled = false, itemWriteStrategy = ItemWriteStrategy.ItemAppend)
+    UpsertParameterTest(bulkEnabled = true, itemWriteStrategy = ItemWriteStrategy.ItemOverwrite, initialBatchSize = None),
+    UpsertParameterTest(bulkEnabled = true, itemWriteStrategy = ItemWriteStrategy.ItemOverwrite, initialBatchSize = Some(1)),
+    UpsertParameterTest(bulkEnabled = false, itemWriteStrategy = ItemWriteStrategy.ItemOverwrite, initialBatchSize = None),
+    UpsertParameterTest(bulkEnabled = false, itemWriteStrategy = ItemWriteStrategy.ItemAppend, initialBatchSize = None)
   )
 
-  for (UpsertParameterTest(bulkEnabled, itemWriteStrategy, hasId) <- upsertParameterTest) {
-    it should s"support upserts with bulkEnabled = $bulkEnabled itemWriteStrategy = $itemWriteStrategy hasId = $hasId" in {
+  for (UpsertParameterTest(bulkEnabled, itemWriteStrategy, hasId, initialBatchSize) <- upsertParameterTest) {
+    it should s"support upserts with bulkEnabled = $bulkEnabled itemWriteStrategy = $itemWriteStrategy hasId = $hasId initialBatchSize = $initialBatchSize" in {
       val cosmosEndpoint = TestConfigurations.HOST
       val cosmosMasterKey = TestConfigurations.MASTER_KEY
 
-      val cfg = Map("spark.cosmos.accountEndpoint" -> cosmosEndpoint,
-        "spark.cosmos.accountKey" -> cosmosMasterKey,
-        "spark.cosmos.database" -> cosmosDatabase,
-        "spark.cosmos.container" -> cosmosContainer,
-        "spark.cosmos.serialization.inclusionMode" -> "NonDefault"
-      )
+      val cfg = {
+
+        initialBatchSize match {
+          case Some(customInitialBatchSize) =>
+            Map(
+              "spark.cosmos.accountEndpoint" -> cosmosEndpoint,
+              "spark.cosmos.accountKey" -> cosmosMasterKey,
+              "spark.cosmos.database" -> cosmosDatabase,
+              "spark.cosmos.container" -> cosmosContainer,
+              "spark.cosmos.serialization.inclusionMode" -> "NonDefault",
+              "spark.cosmos.write.bulk.initialBatchSize" -> customInitialBatchSize.toString,
+            )
+          case None =>
+            Map (
+              "spark.cosmos.accountEndpoint" -> cosmosEndpoint,
+              "spark.cosmos.accountKey" -> cosmosMasterKey,
+              "spark.cosmos.database" -> cosmosDatabase,
+              "spark.cosmos.container" -> cosmosContainer,
+              "spark.cosmos.serialization.inclusionMode" -> "NonDefault"
+            )
+        }
+      }
 
       val cfgOverwrite = Map("spark.cosmos.accountEndpoint" -> cosmosEndpoint,
         "spark.cosmos.accountKey" -> cosmosMasterKey,
@@ -488,6 +504,162 @@ class SparkE2EWriteITest
       quark.get("id").asText() shouldEqual "Quark"
       quark.get("car").get("carType").asText() shouldEqual "X5"
     }
+  }
+
+  private case class PatchBulkUpdateParameterTest(bulkEnabled: Boolean, patchColumnConfigString: String)
+
+  private var patchBulkUpdateParameterTest = Seq(
+      PatchBulkUpdateParameterTest(bulkEnabled = true, "[col(color), col(spin).path(/spin)]"),
+      PatchBulkUpdateParameterTest(bulkEnabled = false, "[col(color), col(spin).path(/spin)]"),
+  )
+
+  for (PatchBulkUpdateParameterTest(bulkEnabled, patchColumnConfigString) <- patchBulkUpdateParameterTest) {
+      it should s"support patch bulk update with bulkEnabled = $bulkEnabled columnConfigString = $patchColumnConfigString" in {
+          val cosmosEndpoint = TestConfigurations.HOST
+          val cosmosMasterKey = TestConfigurations.MASTER_KEY
+
+          val cfg = Map("spark.cosmos.accountEndpoint" -> cosmosEndpoint,
+              "spark.cosmos.accountKey" -> cosmosMasterKey,
+              "spark.cosmos.database" -> cosmosDatabase,
+              "spark.cosmos.container" -> cosmosContainer,
+              "spark.cosmos.serialization.inclusionMode" -> "NonDefault"
+          )
+
+          val cfgPatch = Map("spark.cosmos.accountEndpoint" -> cosmosEndpoint,
+              "spark.cosmos.accountKey" -> cosmosMasterKey,
+              "spark.cosmos.database" -> cosmosDatabase,
+              "spark.cosmos.container" -> cosmosContainer,
+              "spark.cosmos.write.strategy" -> ItemWriteStrategy.ItemBulkUpdate.toString,
+              "spark.cosmos.write.bulk.enabled" -> bulkEnabled.toString,
+              "spark.cosmos.write.patch.columnConfigs" -> patchColumnConfigString
+          )
+
+          val newSpark = getSpark
+
+          // scalastyle:off underscore.import
+          // scalastyle:off import.grouping
+          import spark.implicits._
+          val spark = newSpark
+          // scalastyle:on underscore.import
+          // scalastyle:on import.grouping
+
+          val df = Seq(
+              ("Quark", "Quark", "Red", 1.0 / 2, "")
+          ).toDF("particle name", "id", "color", "spin", "empty")
+
+          df.write.format("cosmos.oltp").mode("Append").options(cfg).save()
+
+          val patchBulkUpdateDf = Seq(
+              ("Quark", "green", 0.03)
+          ).toDF("id", "color", "spin")
+
+          patchBulkUpdateDf.write.format("cosmos.oltp").mode("Append").options(cfgPatch).save()
+
+          // verify data is written
+          // wait for a second to allow replication is completed.
+          Thread.sleep(1000)
+
+          // the item with the same id/pk will be persisted based on the upsert config
+          val quarks = queryItems("SELECT * FROM r where r.id = 'Quark'").toArray
+          quarks should have size 1
+
+          val quark = quarks(0)
+          quark.get("particle name").asText() shouldEqual "Quark"
+          quark.get("id").asText() shouldEqual "Quark"
+          quark.get("color").asText() shouldEqual "green"
+          quark.get("spin").asDouble() shouldEqual 0.03
+      }
+  }
+
+  patchBulkUpdateParameterTest = Seq(
+      PatchBulkUpdateParameterTest(bulkEnabled = true, "[col(car)]"),
+      PatchBulkUpdateParameterTest(bulkEnabled = false, "[col(car)]"),
+      PatchBulkUpdateParameterTest(bulkEnabled = true, "[col(car).rawJson]"),
+      PatchBulkUpdateParameterTest(bulkEnabled = false, "[col(car).rawJson]")
+  )
+
+  for (PatchBulkUpdateParameterTest(bulkEnabled, patchColumnConfigString) <- patchBulkUpdateParameterTest) {
+      it should s"support patch with bulkEnabled = $bulkEnabled columnConfigString = $patchColumnConfigString" in {
+          val cosmosEndpoint = TestConfigurations.HOST
+          val cosmosMasterKey = TestConfigurations.MASTER_KEY
+
+          val cfg = Map("spark.cosmos.accountEndpoint" -> cosmosEndpoint,
+              "spark.cosmos.accountKey" -> cosmosMasterKey,
+              "spark.cosmos.database" -> cosmosDatabase,
+              "spark.cosmos.container" -> cosmosContainer,
+              "spark.cosmos.serialization.inclusionMode" -> "NonDefault"
+          )
+
+          val cfgPatch = Map("spark.cosmos.accountEndpoint" -> cosmosEndpoint,
+              "spark.cosmos.accountKey" -> cosmosMasterKey,
+              "spark.cosmos.database" -> cosmosDatabase,
+              "spark.cosmos.container" -> cosmosContainer,
+              "spark.cosmos.write.strategy" -> ItemWriteStrategy.ItemBulkUpdate.toString,
+              "spark.cosmos.write.bulk.enabled" -> bulkEnabled.toString,
+              "spark.cosmos.write.bulkUpdate.columnConfigs" -> patchColumnConfigString
+          )
+
+          val newSpark = getSpark
+
+          // scalastyle:off underscore.import
+          // scalastyle:off import.grouping
+          import spark.implicits._
+          val spark = newSpark
+          // scalastyle:on underscore.import
+          // scalastyle:on import.grouping
+
+          val dfWithJson = Seq(
+              ("Quark", "Quark", "Red", 1.0 / 2, "", "{ \"manufacturer\": \"BMW\", \"carType\": \"X3\" }")
+          ).toDF("particle name", "id", "color", "spin", "empty", "childNodeJson")
+
+          val df = dfWithJson
+              .withColumn("car", from_json(col("childNodeJson"), StructType(Array(StructField("manufacturer", StringType, true), StructField("carType", StringType, true)))))
+              .drop("childNodeJson")
+          df.show(false)
+          df.write.format("cosmos.oltp").mode("Append").options(cfg).save()
+
+          // verify data is written
+          // wait for a second to allow replication is completed.
+          Thread.sleep(1000)
+
+          // the item with the same id/pk will be persisted based on the upsert config
+          var quarks = queryItems("SELECT * FROM r where r.id = 'Quark'").toArray
+          quarks should have size 1
+
+          var quark = quarks(0)
+          quark.get("particle name").asText() shouldEqual "Quark"
+          quark.get("id").asText() shouldEqual "Quark"
+          quark.get("car").get("carType").asText() shouldEqual "X3"
+
+          val patchDf = if (patchColumnConfigString.endsWith(".rawJson]")) {
+              Seq(("Quark", "{ \"manufacturer\": \"BMW\", \"carType\": \"X5\" }"))
+                  .toDF("id", "car")
+          } else {
+              Seq(("Quark", "{ \"manufacturer\": \"BMW\", \"carType\": \"X5\" }"))
+                  .toDF("id", "childNodeJson")
+                  .withColumn("car", from_json(col("childNodeJson"), StructType(Array(StructField("manufacturer", StringType, true), StructField("carType", StringType, true)))))
+                  .drop("childNodeJson")
+          }
+
+          logInfo(s"Schema of patchDf: ${patchDf.schema}")
+
+          patchDf.write.format("cosmos.oltp").mode("Append").options(cfgPatch).save()
+
+          // verify data is written
+          // wait for a second to allow replication is completed.
+          Thread.sleep(1000)
+
+          // the item with the same id/pk will be persisted based on the upsert config
+          quarks = queryItems("SELECT * FROM r where r.id = 'Quark'").toArray
+          quarks should have size 1
+
+          logInfo(s"JSON returned from query: ${quarks(0)}")
+
+          quark = quarks(0)
+          quark.get("particle name").asText() shouldEqual "Quark"
+          quark.get("id").asText() shouldEqual "Quark"
+          quark.get("car").get("carType").asText() shouldEqual "X5"
+      }
   }
   //scalastyle:on magic.number
   //scalastyle:on multiple.string.literals