+++ sdk/formrecognizer/azure-ai-formrecognizer/src/main/java/com/azure/ai/formrecognizer/documentanalysis/implementation/models/AnalyzeDocumentRequest.java
@@ -12,13 +12,13 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 @Fluent
 public final class AnalyzeDocumentRequest {
     /*
-     * Document URL to analyze
+     * Document URL to analyze. Either urlSource or base64Source must be specified.
      */
     @JsonProperty(value = "urlSource")
     private String urlSource;
 
     /*
-     * Base64 encoding of the document to analyze
+     * Base64 encoding of the document to analyze. Either urlSource or base64Source must be specified.
      */
     @JsonProperty(value = "base64Source")
     private byte[] base64Source;
@@ -27,7 +27,7 @@ public final class AnalyzeDocumentRequest {
     public AnalyzeDocumentRequest() {}
 
     /**
-     * Get the urlSource property: Document URL to analyze.
+     * Get the urlSource property: Document URL to analyze. Either urlSource or base64Source must be specified.
      *
      * @return the urlSource value.
      */
@@ -36,7 +36,7 @@ public final class AnalyzeDocumentRequest {
     }
 
     /**
-     * Set the urlSource property: Document URL to analyze.
+     * Set the urlSource property: Document URL to analyze. Either urlSource or base64Source must be specified.
      *
      * @param urlSource the urlSource value to set.
      * @return the AnalyzeDocumentRequest object itself.
@@ -47,7 +47,8 @@ public final class AnalyzeDocumentRequest {
     }
 
     /**
-     * Get the base64Source property: Base64 encoding of the document to analyze.
+     * Get the base64Source property: Base64 encoding of the document to analyze. Either urlSource or base64Source must
+     * be specified.
      *
      * @return the base64Source value.
      */
@@ -56,7 +57,8 @@ public final class AnalyzeDocumentRequest {
     }
 
     /**
-     * Set the base64Source property: Base64 encoding of the document to analyze.
+     * Set the base64Source property: Base64 encoding of the document to analyze. Either urlSource or base64Source must
+     * be specified.
      *
      * @param base64Source the base64Source value to set.
      * @return the AnalyzeDocumentRequest object itself.