+++ sdk/formrecognizer/azure-ai-formrecognizer/src/main/java/com/azure/ai/formrecognizer/documentanalysis/models/TrainingDataContentSource.java
@@ -1,0 +1,10 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+package com.azure.ai.formrecognizer.documentanalysis.models;
+
+/**
+ * Class representing the source of the training content.
+ */
+public abstract class TrainingDataContentSource {
+}