+++ sdk/identity/azure-identity/src/main/java/com/azure/identity/implementation/IdentityClientOptions.java
@@ -53,7 +53,6 @@ public final class IdentityClientOptions implements Cloneable {
     private boolean includeX5c;
     private AuthenticationRecord authenticationRecord;
     private TokenCachePersistenceOptions tokenCachePersistenceOptions;
-    private boolean cp1Disabled;
     private RegionalAuthority regionalAuthority;
     private UserAssertion userAssertion;
     private boolean multiTenantAuthDisabled;
@@ -346,15 +345,6 @@ public final class IdentityClientOptions implements Cloneable {
     }
 
     /**
-     * Check whether CP1 client capability should be disabled.
-     *
-     * @return the status indicating if CP1 client capability should be disabled.
-     */
-    public boolean isCp1Disabled() {
-        return this.cp1Disabled;
-    }
-
-    /**
      * Gets the regional authority, or null if regional authority should not be used.
      * @return the regional authority value if specified
      */
@@ -605,11 +595,6 @@ public final class IdentityClientOptions implements Cloneable {
         return this.perCallPolicies;
     }
 
-    IdentityClientOptions setCp1Disabled(boolean cp1Disabled) {
-        this.cp1Disabled = cp1Disabled;
-        return this;
-    }
-
     IdentityClientOptions setMultiTenantAuthDisabled(boolean multiTenantAuthDisabled) {
         this.multiTenantAuthDisabled = multiTenantAuthDisabled;
         return this;
@@ -690,7 +675,6 @@ public final class IdentityClientOptions implements Cloneable {
         imdsAuthorityHost = configuration.get(AZURE_POD_IDENTITY_AUTHORITY_HOST,
             IdentityConstants.DEFAULT_IMDS_ENDPOINT);
         ValidationUtil.validateAuthHost(authorityHost, LOGGER);
-        cp1Disabled = configuration.get(Configuration.PROPERTY_AZURE_IDENTITY_DISABLE_CP1, false);
         multiTenantAuthDisabled = configuration
             .get(AZURE_IDENTITY_DISABLE_MULTI_TENANT_AUTH, false);
     }
@@ -747,7 +731,6 @@ public final class IdentityClientOptions implements Cloneable {
             .setIntelliJKeePassDatabasePath(this.keePassDatabasePath)
             .setAuthorityHost(this.authorityHost)
             .setImdsAuthorityHost(this.imdsAuthorityHost)
-            .setCp1Disabled(this.cp1Disabled)
             .setMultiTenantAuthDisabled(this.multiTenantAuthDisabled)
             .setUserAssertion(this.userAssertion)
             .setConfigurationStore(this.configuration)