+++ sdk/search/azure-search-documents/CHANGELOG.md
@@ -10,6 +10,17 @@
 
 ### Other Changes
 
+## 11.5.9 (2023-07-25)
+
+### Other Changes
+
+#### Dependency Updates
+
+- Upgraded `azure-core-http-netty` from `1.13.4` to version `1.13.5`.
+- Upgraded `azure-core-serializer-json-jackson` from `1.4.1` to version `1.4.2`.
+- Upgraded `azure-core` from `1.40.0` to version `1.41.0`.
+
+
 ## 11.6.0-beta.7 (2023-07-11)
 
 ### Features Added